reorg into library crate
