+++ src/main.rs
@@ -1,7 +1,7 @@
 use std::env;
-use std::error::Error;
-use std::fs;
 use std::process;
+
+use bb_rust_grep::Config;
 
 fn main() {
     let args: Vec<String> = env::args().collect();
@@ -14,33 +14,8 @@ fn main() {
     println!("Searching for {}", config.query);
     println!("In file {}", config.file_path);
 
-    if let Err(e) = run(config) {
+    if let Err(e) = bb_rust_grep::run(config) {
         println!("Application error: {e}");
         process::exit(1);
-    }
-}
-
-fn run(config: Config) -> Result<(), Box<dyn Error>> {
-    let contents = fs::read_to_string(config.file_path)?;
-
-    println!("With text:\n{contents}");
-
-    Ok(())
-}
-
-struct Config {
-    query: String,
-    file_path: String,
-}
-
-impl Config {
-    fn build(args: &[String]) -> Result<Config, &'static str> {
-        if args.len() < 3 {
-            return Err("Insufficient number of arguments");
-        }
-        let query = args[1].clone();
-        let file_path = args[2].clone();
-
-        Ok(Config { query, file_path })
     }
 }
